+++ src/main/java/net/raphimc/noteblocktool/elements/drag/DragTable.java
@@ -43,7 +43,7 @@ public class DragTable extends JTable {
         List<RowSorter.SortKey> sortKeys = new ArrayList<>();
         for (int i = 0; i < 6; i++) sortKeys.add(new RowSorter.SortKey(i, SortOrder.UNSORTED));
         sorter.setSortKeys(sortKeys);
-        sorter.setComparator(4, Comparator.comparingLong(o -> (long) o));
+        sorter.setComparator(4, Comparator.comparingInt(o -> (int) o));
         sorter.setComparator(5, Comparator.comparingDouble(o -> (float) o));
         this.setRowSorter(sorter);
     }
